feat: Support PureJsonView component
